+++ paddle_billing/Entities/Event.py
@@ -7,6 +7,7 @@ from paddle_billing.Entities.Entity import Entity
 from paddle_billing.Entities.Events import EventTypeName
 
 from paddle_billing.Notifications.Entities.Entity import Entity as NotificationEntity
+from paddle_billing.Notifications.Entities.UndefinedEntity import UndefinedEntity
 
 
 @dataclass
@@ -14,7 +15,7 @@ class Event(Entity, ABC):
     event_id: str
     event_type: EventTypeName
     occurred_at: datetime
-    data: NotificationEntity | dict
+    data: NotificationEntity | UndefinedEntity
 
     @staticmethod
     def from_dict(data: dict) -> Event: