+++ tests/Unit/Entities/Notifications/test_NotificationEvent.py
@@ -6,6 +6,7 @@ from paddle_billing.Entities.Notifications import NotificationEvent
 
 from paddle_billing.Notifications.Entities.Subscription import Subscription
 from paddle_billing.Notifications.Entities.SubscriptionCreated import SubscriptionCreated
+from paddle_billing.Notifications.Entities.UndefinedEntity import UndefinedEntity
 
 from tests.Utils.ReadsFixture import ReadsFixtures
 
@@ -127,6 +128,8 @@ class TestNotificationEvent:
         assert notification_event.event_id == "evt_01h8bzakzx3hm2fmen703n5q45"
         assert notification_event.event_type == event_type
         assert notification_event.occurred_at.isoformat() == "2023-08-21T11:57:47.390028+00:00"
+        assert isinstance(notification_event.data.to_dict(), dict)
+        assert notification_event.data.to_dict()["id"] is not None
 
     def test_subscription_created_notification_event_transaction_id(self):
         notification_event = NotificationEvent.from_dict(
@@ -141,6 +144,8 @@ class TestNotificationEvent:
 
         assert isinstance(notification_event.data, SubscriptionCreated)
         assert notification_event.data.transaction_id == "txn_01hv8wptq8987qeep44cyrewp9"
+        assert isinstance(notification_event.data.to_dict(), dict)
+        assert notification_event.data.to_dict()["transaction_id"] == "txn_01hv8wptq8987qeep44cyrewp9"
 
     @mark.parametrize(
         "event_type",
@@ -198,5 +203,5 @@ class TestNotificationEvent:
         assert notification_event.event_id == "evt_01h8bzakzx3hm2fmen703n5q45"
         assert notification_event.event_type == "some_unknown_entity.created"
         assert notification_event.occurred_at.isoformat() == "2023-08-21T11:57:47.390028+00:00"
-        assert isinstance(notification_event.data, dict)
-        assert notification_event.data["id"] == "add_01hv8gq3318ktkfengj2r75gfx"
+        assert isinstance(notification_event.data, UndefinedEntity)
+        assert notification_event.data.to_dict()["id"] == "add_01hv8gq3318ktkfengj2r75gfx"